Convert a vtreat tstep into an rquery operation tree.

Convert a vtreat tstep into an rquery operation tree.

as_rquery(tstep, ...)

Arguments

tstep
vtreat treatment plan.

...
not used, force any later arguments to bind to names.

Value

list(optree_generator (ordered list of functions), temp_tables (named list of tables))

Examples

if(requireNamespace("rquery", quietly = TRUE)) {
   dTrainC <- data.frame(x= c('a', 'a', 'a', 'b' ,NA , 'b'),
                         z= c(1, 2, NA, 4, 5, 6),
                         y= c(FALSE, FALSE, TRUE, FALSE, TRUE, TRUE),
                         stringsAsFactors = FALSE)
   dTrainC$id <- seq_len(nrow(dTrainC))
   treatmentsC <- designTreatmentsC(dTrainC, c("x", "z"), 'y', TRUE)
   rqplan <- as_rquery(treatmentsC)
   ops <- flatten_fn_list(rquery::local_td(dTrainC), rqplan$optree_generators)
   cat(format(ops))
   if(requireNamespace("rqdatatable", quietly = TRUE)) {
      treated <- rqdatatable::ex_data_table(ops, tables = rqplan$tables)
      print(treated[])
   }
   if(requireNamespace("DBI", quietly = TRUE) &&
      requireNamespace("DBI", quietly = TRUE)) {
      db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
      source_data <- rquery::rq_copy_to(db, "dTrainC", dTrainC,
                               overwrite = TRUE, temporary = TRUE)

      rest <- materialize_treated(db, rqplan, source_data, "dTreatedC",
                                  extracols = "id",
                                  print_sql = FALSE)
      resd <- DBI::dbReadTable(db, rest$table_name)
      print(resd)

      rquery::rq_remove_table(db, source_data$table_name)
      rquery::rq_remove_table(db, rest$table_name)
      DBI::dbDisconnect(db)
   }
}

#> [1] "vtreat 1.2.0 inspecting inputs Fri Jun 15 11:22:10 2018"
#> [1] "designing treatments Fri Jun 15 11:22:10 2018"
#> [1] " have initial level statistics Fri Jun 15 11:22:10 2018"
#> [1] "design var x Fri Jun 15 11:22:10 2018"
#> [1] "design var z Fri Jun 15 11:22:10 2018"
#> [1] " scoring treatments Fri Jun 15 11:22:10 2018"
#> [1] "have treatment plan Fri Jun 15 11:22:10 2018"
#> [1] "rescoring complex variables Fri Jun 15 11:22:10 2018"
#> [1] "done rescoring complex variables Fri Jun 15 11:22:10 2018"

#> Warning: vtreat::as_rquery not yet implemented for  vtreat 'Prevalence Code'('x'(character,character)->character->'x_catP') , class vtreatment

#> table('dTrainC'; 
#>   x,
#>   z,
#>   y,
#>   id) %.>%
#>  extend(.,
#>   x_lev_NA := ifelse(is.na(x), 1, 0),
#>   x_lev_x.a := ifelse(is.na(x), 0, ifelse(x == "a", 1, 0)),
#>   x_lev_x.b := ifelse(is.na(x), 0, ifelse(x == "b", 1, 0)),
#>   z_clean := ifelse(is.na(z), 3.6, z),
#>   z_isBAD := ifelse(is.na(z), 1, 0)) %.>%
#>  natural_join(.,
#>   table('vtreat_tmp_03414039267197056602_0000000000'; 
#>     x,
#>     x_catB),
#>   j= LEFT, by= x) %.>%
#>  extend(.,
#>   x_catB := ifelse(is.na(x), 9.21044036697607, ifelse(is.na(x_catB), 0, x_catB)))
#>       x id     y  z x_lev_NA x_lev_x.a x_lev_x.b z_clean z_isBAD     x_catB
#> 1: <NA>  5  TRUE  5        1         0         0     5.0       0  9.2104404
#> 2:    a  1 FALSE  1        0         1         0     1.0       0 -0.6930972
#> 3:    a  2 FALSE  2        0         1         0     2.0       0 -0.6930972
#> 4:    a  3  TRUE NA        0         1         0     3.6       1 -0.6930972
#> 5:    b  4 FALSE  4        0         0         1     4.0       0  0.0000000
#> 6:    b  6  TRUE  6        0         0         1     6.0       0  0.0000000
#>   id     x_catB x_lev_NA x_lev_x.a x_lev_x.b y z_clean z_isBAD
#> 1  1 -0.6930972        0         1         0 0     1.0       0
#> 2  2 -0.6930972        0         1         0 0     2.0       0
#> 3  3 -0.6930972        0         1         0 1     3.6       1
#> 4  4  0.0000000        0         0         1 0     4.0       0
#> 5  5  9.2104404        1         0         0 1     5.0       0
#> 6  6  0.0000000        0         0         1 1     6.0       0