Build set carve-up for out-of sample evaluation.

Return a carve-up of seq_len(nRows). Very useful for any sort of nested model situation (such as data prep, stacking, or super-learning).

buildEvalSets(
  nRows,
  ...,
  dframe = NULL,
  y = NULL,
  splitFunction = NULL,
  nSplits = 3
)

Arguments

nRows
scalar, >=1 number of rows to sample from.

...
no additional arguments, declared to forced named binding of later arguments.

dframe
(optional) original data.frame, passed to user splitFunction.

y
(optional) numeric vector, outcome variable (possibly to stratify on), passed to user splitFunction.

splitFunction
(optional) function taking arguments nSplits,nRows,dframe, and y; returning a user desired split.

nSplits
integer, target number of splits.

Value

list of lists where the app portion of the sub-lists is a disjoint carve-up of seq_len(nRows) and each list as a train portion disjoint from app.

Details

Also sets attribute "splitmethod" on return value that describes how the split was performed. attr(returnValue,'splitmethod') is one of: 'notsplit' (data was not split; corner cases like single row data sets), 'oneway' (leave one out holdout), 'kwaycross' (a simple partition), 'userfunction' (user supplied function was actually used), or a user specified attribute. Any user desired properties (such as stratification on y, or preservation of groups designated by original data row numbers) may not apply unless you see that 'userfunction' has been used.

The intent is the user splitFunction only needs to handle "easy cases" and maintain user invariants. If the user splitFunction returns NULL, throws, or returns an unacceptable carve-up then vtreat::buildEvalSets returns its own eval set plan. The signature of splitFunction should be splitFunction(nRows,nSplits,dframe,y) where nSplits is the number of pieces we want in the carve-up, nRows is the number of rows to split, dframe is the original dataframe (useful for any group control variables), and y is a numeric vector representing outcome (useful for outcome stratification).

Note that buildEvalSets may not always return a partition (such as one row dataframes), or if the user split function chooses to make rows eligible for application a different number of times.

See also

Examples

# use
buildEvalSets(200)

#> [[1]]
#> [[1]]$train
#>   [1]   1   4   5   6   9  10  12  14  15  16  17  19  20  21  22  23  24  25
#>  [19]  26  27  28  32  34  35  36  37  39  40  42  43  46  47  49  50  51  52
#>  [37]  53  54  55  56  58  59  61  63  64  65  66  67  69  74  78  80  81  82
#>  [55]  84  85  86  87  88  89  90  92  93  94  95  97  99 100 101 103 105 106
#>  [73] 107 108 110 111 113 114 116 118 120 121 126 127 128 129 130 131 132 133
#>  [91] 136 137 138 139 140 141 143 144 147 150 151 152 153 154 155 157 158 159
#> [109] 160 161 164 165 167 168 169 170 171 172 173 174 175 176 179 181 182 183
#> [127] 185 187 188 189 191 192 196 198
#> 
#> [[1]]$app
#>  [1]  38 134   2 156  41 102  60 115  13  77  11   7  73 109  45 186 193 145   3
#> [20] 194 190  44  29 184 177 200  83  57  31 199 162 122 149  79 117  75 180 142
#> [39]  72  71  33  96 195 197 123 112  68 163 124  30  62  91   8  18  70 119 178
#> [58] 148 146  48 166 125 135  76 104  98
#> 
#> 
#> [[2]]
#> [[2]]$train
#>   [1]   2   3   4   5   7   8  10  11  13  15  16  17  18  20  21  22  23  26
#>  [19]  27  28  29  30  31  33  34  35  36  37  38  39  40  41  44  45  47  48
#>  [37]  49  52  53  55  57  58  59  60  61  62  65  66  67  68  69  70  71  72
#>  [55]  73  75  76  77  78  79  80  83  86  87  91  94  95  96  98 102 104 105
#>  [73] 107 109 110 112 115 116 117 118 119 122 123 124 125 128 129 130 131 133
#>  [91] 134 135 137 138 142 145 146 147 148 149 153 154 155 156 157 158 160 161
#> [109] 162 163 166 168 169 170 176 177 178 180 182 183 184 185 186 188 190 191
#> [127] 193 194 195 197 198 199 200
#> 
#> [[2]]$app
#>  [1] 106 196 150  64  12  42 103  14 101  97  32 172 151  51  25 164  82 167 174
#> [20] 165 139 113  85  74 121 152 108  56 175  93  24  89 120 132  63 114  54 143
#> [39] 144 126 181 192  43 136  90 111  92  50 140  19 179   6  88 159  81  46 100
#> [58]  84   1 171 189   9  99 173 127 187 141
#> 
#> 
#> [[3]]
#> [[3]]$train
#>   [1]   1   2   3   6   7   8   9  11  12  13  14  18  19  24  25  29  30  31
#>  [19]  32  33  38  41  42  43  44  45  46  48  50  51  54  56  57  60  62  63
#>  [37]  64  68  70  71  72  73  74  75  76  77  79  81  82  83  84  85  88  89
#>  [55]  90  91  92  93  96  97  98  99 100 101 102 103 104 106 108 109 111 112
#>  [73] 113 114 115 117 119 120 121 122 123 124 125 126 127 132 134 135 136 139
#>  [91] 140 141 142 143 144 145 146 148 149 150 151 152 156 159 162 163 164 165
#> [109] 166 167 171 172 173 174 175 177 178 179 180 181 184 186 187 189 190 192
#> [127] 193 194 195 196 197 199 200
#> 
#> [[3]]$app
#>  [1] 131  65 147  40 182  66  26   5  17 161  55  36  34 105  53   4  94 128  39
#> [20] 188 185  15  58 183  69  37 154  52  87 118 169  10 107  49  23  59 138 130
#> [39]  16 129 191 160  78 170 158  95  47 133 155  20 137 110  22  86  28  80 198
#> [58] 157  67  35 116 176 168 153  27  61  21
#> 
#> 
#> attr(,"splitmethod")
#> [1] "kwaycross"

# longer example
# helper fns
# fit models using experiment plan to estimate out of sample behavior
fitModelAndApply <- function(trainData,applicaitonData) {
   model <- lm(y~x,data=trainData)
   predict(model,newdata=applicaitonData)
}
simulateOutOfSampleTrainEval <- function(d,fitApplyFn) {
   eSets <- buildEvalSets(nrow(d))
   evals <- lapply(eSets,
      function(ei) { fitApplyFn(d[ei$train,],d[ei$app,]) })
   pred <- numeric(nrow(d))
   for(eii in seq_len(length(eSets))) {
     pred[eSets[[eii]]$app] <- evals[[eii]]
   }
   pred
}

# run the experiment
set.seed(2352356)
# example data
d <- data.frame(x=rnorm(5),y=rnorm(5),
        outOfSampleEst=NA,inSampleEst=NA)

# fit model on all data
d$inSampleEst <- fitModelAndApply(d,d)
# compute in-sample R^2 (above zero, falsely shows a 
#   relation until we adjust for degrees of freedom)
1-sum((d$y-d$inSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] 0.4193942

d$outOfSampleEst <- simulateOutOfSampleTrainEval(d,fitModelAndApply)
# compute out-sample R^2 (not positive, 
#  evidence of no relation)
1-sum((d$y-d$outOfSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] -3.873148