Design a simple treatment plan to indicate missingingness and perform simple imputation.

Design a simple treatment plan to indicate missingingness and perform simple imputation.

design_missingness_treatment(
  dframe,
  ...,
  varlist = colnames(dframe),
  invalid_mark = "_invalid_",
  drop_constant_columns = FALSE,
  missingness_imputation = NULL,
  imputation_map = NULL
)

Arguments

dframe
data.frame to drive design.

...
not used, forces later arguments to bind by name.

varlist
character, names of columns to process.

invalid_mark
character, name to use for NA levels and novel levels.

drop_constant_columns
logical, if TRUE drop columns that do not vary from the treatment plan.

missingness_imputation
function of signature f(values: numeric), simple missing value imputer.

imputation_map
map from column names to functions of signature f(values: numeric), simple missing value imputers.

Value

simple treatment plan.

See also

Examples

d <- wrapr::build_frame(
  "x1", "x2", "x3" |
  1   , 4   , "A"  |
  NA  , 5   , "B"  |
  3   , 6   , NA   )

plan <- design_missingness_treatment(d)
prepare(plan, d)

#>   x1 x1_isBAD x2        x3
#> 1  1        0  4         A
#> 2  2        1  5         B
#> 3  3        0  6 _invalid_

prepare(plan, data.frame(x1=NA, x2=NA, x3="E"))

#>   x1 x1_isBAD x2        x3
#> 1  2        1  5 _invalid_