check if appPlan is a good carve-up of 1:nRows into nSplits groups

check if appPlan is a good carve-up of 1:nRows into nSplits groups

problemAppPlan(nRows, nSplits, appPlan, strictCheck)

Arguments

nRows

number of rows to carve-up

nSplits

number of sets to carve-up into

appPlan

carve-up to critique

strictCheck

logical, if true expect application data to be a carve-up and training data to be a maximal partition and to match nSplits.

Value

problem with carve-up (null if good)

See also

Examples

plan <- kWayStratifiedY(3,2,NULL,NULL)
problemAppPlan(3,3,plan,TRUE)
#> [1] "didn't get requested number of groups in appPlan"