Apply a treatment plan using rqdatatable.

Note: does not treat map NaN or +-Infinity. This function is only for timings and demonstration, not for production use.

rqdatatable_prepare(
  rqplan,
  data_source,
  ...,
  partition_column = NULL,
  parallelCluster = NULL,
  use_parallel = use_parallel,
  extracols = NULL,
  non_join_mapping = FALSE,
  print_rquery = FALSE,
  env = parent.frame()
)

Arguments

rqplan
an query plan produced by as_rquery_plan().

data_source
a data.frame.

...
force later arguments to bind by name.

partition_column
character name of column to partition work by.

parallelCluster
a cluster object, created by package parallel or by package snow. If NULL, use the registered default cluster.

use_parallel
logical, if TRUE use parallel cluster (when available).

extracols
extra columns to copy.

non_join_mapping
logical, if TRUE use non-join based column mapping.

print_rquery
logical, if TRUE print the rquery ops.

env
environment to work in.

Value

treated data.

See also