Build set carve-up for out-of sample evaluation.

Return a carve-up of seq_len(nRows). Very useful for any sort of nested model situation (such as data prep, stacking, or super-learning).

buildEvalSets(
  nRows,
  ...,
  dframe = NULL,
  y = NULL,
  splitFunction = NULL,
  nSplits = 3
)

Arguments

nRows

scalar, >=1 number of rows to sample from.

...

no additional arguments, declared to forced named binding of later arguments.

dframe

(optional) original data.frame, passed to user splitFunction.

y

(optional) numeric vector, outcome variable (possibly to stratify on), passed to user splitFunction.

splitFunction

(optional) function taking arguments nSplits,nRows,dframe, and y; returning a user desired split.

nSplits

integer, target number of splits.

Value

list of lists where the app portion of the sub-lists is a disjoint carve-up of seq_len(nRows) and each list as a train portion disjoint from app.

Details

Also sets attribute "splitmethod" on return value that describes how the split was performed. attr(returnValue,'splitmethod') is one of: 'notsplit' (data was not split; corner cases like single row data sets), 'oneway' (leave one out holdout), 'kwaycross' (a simple partition), 'userfunction' (user supplied function was actually used), or a user specified attribute. Any user desired properties (such as stratification on y, or preservation of groups designated by original data row numbers) may not apply unless you see that 'userfunction' has been used.

The intent is the user splitFunction only needs to handle "easy cases" and maintain user invariants. If the user splitFunction returns NULL, throws, or returns an unacceptable carve-up then vtreat::buildEvalSets returns its own eval set plan. The signature of splitFunction should be splitFunction(nRows,nSplits,dframe,y) where nSplits is the number of pieces we want in the carve-up, nRows is the number of rows to split, dframe is the original dataframe (useful for any group control variables), and y is a numeric vector representing outcome (useful for outcome stratification).

Note that buildEvalSets may not always return a partition (such as one row dataframes), or if the user split function chooses to make rows eligible for application a different number of times.

See also

Examples

# use
buildEvalSets(200)
#> [[1]]
#> [[1]]$train
#>   [1]   1   2   4   7   8  11  12  14  15  16  17  18  19  20  21  22  24  25
#>  [19]  26  27  29  32  33  36  37  39  40  41  42  43  44  46  47  48  49  50
#>  [37]  51  53  54  56  59  60  62  65  66  67  69  71  72  73  75  76  77  78
#>  [55]  81  82  83  84  85  88  89  91  93  94  95  96  97  98  99 100 101 102
#>  [73] 103 104 105 106 108 109 112 114 115 116 120 122 123 125 127 128 133 134
#>  [91] 136 137 138 140 141 144 145 148 150 151 152 155 159 160 161 162 163 164
#> [109] 165 167 169 170 172 173 174 175 176 177 178 181 182 183 184 186 187 190
#> [127] 191 193 195 196 197 198 199 200
#> 
#> [[1]]$app
#>  [1]  30 130 132  61  70 131 157 139  10  28 124 153  45  90   3 121   5  13 129
#> [20]  31  58 119  63 171 158  80  64 189 113   6   9 147 107  23 156  34 179 126
#> [39]  86 188  68  52 118  35 135  74  55 117  57 180 146 111 185 149  79 142 110
#> [58] 154  38 143 194 166 192  92 168  87
#> 
#> 
#> [[2]]
#> [[2]]$train
#>   [1]   2   3   4   5   6   7   8   9  10  11  13  15  17  19  20  22  23  24
#>  [19]  28  30  31  32  33  34  35  37  38  41  43  45  47  49  51  52  54  55
#>  [37]  57  58  59  60  61  63  64  67  68  70  71  72  73  74  75  77  78  79
#>  [55]  80  81  82  84  86  87  89  90  92  94  95  98 103 106 107 108 109 110
#>  [73] 111 112 113 114 117 118 119 120 121 122 124 126 129 130 131 132 134 135
#>  [91] 137 138 139 140 142 143 145 146 147 149 151 152 153 154 156 157 158 161
#> [109] 162 164 166 167 168 169 171 176 177 179 180 184 185 186 187 188 189 190
#> [127] 191 192 194 195 196 198 199
#> 
#> [[2]]$app
#>  [1] 197  26 133  53  25 141  36 170  44 155 125  40 150  97   1 165 127  29  27
#> [20] 104 116 115 123  12  42 174  62  76 172  21 173  96 183  56  14  93  85 136
#> [39] 181 105  91  69 178 100 102 182 101 160  18 193  46  88 159 163 128  39  16
#> [58]  65  50 200  48 148 144 175  99  66  83
#> 
#> 
#> [[3]]
#> [[3]]$train
#>   [1]   1   3   5   6   9  10  12  13  14  16  18  21  23  25  26  27  28  29
#>  [19]  30  31  34  35  36  38  39  40  42  44  45  46  48  50  52  53  55  56
#>  [37]  57  58  61  62  63  64  65  66  68  69  70  74  76  79  80  83  85  86
#>  [55]  87  88  90  91  92  93  96  97  99 100 101 102 104 105 107 110 111 113
#>  [73] 115 116 117 118 119 121 123 124 125 126 127 128 129 130 131 132 133 135
#>  [91] 136 139 141 142 143 144 146 147 148 149 150 153 154 155 156 157 158 159
#> [109] 160 163 165 166 168 170 171 172 173 174 175 178 179 180 181 182 183 185
#> [127] 188 189 192 193 194 197 200
#> 
#> [[3]]$app
#>  [1] 199  94  19 108 176  43  81  22  47 122 162 145 140 103 186 187  32  59 167
#> [20] 177  41   8  54 152  78  51 138 164   4  72 106 120  84  33  11  71  98  82
#> [39] 198  95  17 191 112   2 109  24 161  60 114 169  15 137 134 151  77   7  75
#> [58]  89  20  49 196  67 195  37 190 184  73
#> 
#> 
#> attr(,"splitmethod")
#> [1] "kwaycross"

# longer example
# helper fns
# fit models using experiment plan to estimate out of sample behavior
fitModelAndApply <- function(trainData,applicaitonData) {
   model <- lm(y~x,data=trainData)
   predict(model,newdata=applicaitonData)
}
simulateOutOfSampleTrainEval <- function(d,fitApplyFn) {
   eSets <- buildEvalSets(nrow(d))
   evals <- lapply(eSets, 
      function(ei) { fitApplyFn(d[ei$train,],d[ei$app,]) })
   pred <- numeric(nrow(d))
   for(eii in seq_len(length(eSets))) {
     pred[eSets[[eii]]$app] <- evals[[eii]]
   }
   pred
}

# run the experiment
set.seed(2352356)
# example data
d <- data.frame(x=rnorm(5),y=rnorm(5),
        outOfSampleEst=NA,inSampleEst=NA)
        
# fit model on all data
d$inSampleEst <- fitModelAndApply(d,d)
# compute in-sample R^2 (above zero, falsely shows a 
#   relation until we adjust for degrees of freedom)
1-sum((d$y-d$inSampleEst)^2)/sum((d$y-mean(d$y))^2)
#> [1] 0.4193942

d$outOfSampleEst <- simulateOutOfSampleTrainEval(d,fitModelAndApply)
# compute out-sample R^2 (not positive, 
#  evidence of no relation)
1-sum((d$y-d$outOfSampleEst)^2)/sum((d$y-mean(d$y))^2)
#> [1] -3.873148