Build set carve-up for out-of sample evaluation.

Return a carve-up of seq_len(nRows). Very useful for any sort of nested model situation (such as data prep, stacking, or super-learning).

buildEvalSets(nRows, ..., dframe = NULL, y = NULL, splitFunction = NULL,
  nSplits = 3)

Arguments

nRows
scalar, >=1 number of rows to sample from.

...
no additional arguments, declared to forced named binding of later arguments.

dframe
(optional) original data.frame, passed to user splitFunction.

y
(optional) numeric vector, outcome variable (possibly to stratify on), passed to user splitFunction.

splitFunction
(optional) function taking arguments nSplits,nRows,dframe, and y; returning a user desired split.

nSplits
integer, target number of splits.

Value

list of lists where the app portion of the sub-lists is a disjoint carve-up of seq_len(nRows) and each list as a train portion disjoint from app.

Details

Also sets attribute "splitmethod" on return value that describes how the split was performed. attr(returnValue,'splitmethod') is one of: 'notsplit' (data was not split; corner cases like single row data sets), 'oneway' (leave one out holdout), 'kwaycross' (a simple partition), 'userfunction' (user supplied function was actually used), or a user specified attribute. Any user desired properties (such as stratification on y, or preservation of groups designated by original data row numbers) may not apply unless you see that 'userfunction' has been used.

The intent is the user splitFunction only needs to handle "easy cases" and maintain user invariants. If the user splitFunction returns NULL, throws, or returns an unacceptable carve-up then vtreat::buildEvalSets returns its own eval set plan. The signature of splitFunction should be splitFunction(nRows,nSplits,dframe,y) where nSplits is the number of pieces we want in the carve-up, nRows is the number of rows to split, dframe is the original dataframe (useful for any group control variables), and y is a numeric vector representing outcome (useful for outcome stratification).

Note that buildEvalSets may not always return a partition (such as one row dataframes), or if the user split function chooses to make rows eligable for applicaton a different number of times.

See also

Examples

# use
buildEvalSets(200)

#> [[1]]
#> [[1]]$train
#>   [1]   3   4   5   6   7   8   9  10  13  14  15  16  19  21  23  24  25  26
#>  [19]  27  28  30  31  34  35  36  38  39  42  43  44  45  46  49  50  51  52
#>  [37]  53  54  55  56  57  58  60  61  62  63  65  68  74  76  80  81  82  83
#>  [55]  84  85  86  87  88  89  91  93  96  97  98 104 105 106 109 110 111 114
#>  [73] 115 116 117 118 120 121 122 123 124 126 127 128 130 132 139 140 141 142
#>  [91] 143 144 145 147 148 149 150 151 153 154 155 157 158 159 160 161 162 164
#> [109] 165 166 167 168 169 172 173 175 178 179 180 181 182 183 184 186 189 190
#> [127] 191 192 193 194 195 196 197 200
#> 
#> [[1]]$app
#>  [1]  95 185 131 125 187  72  32 138  12  47 101  90  73  75  59 156  40 136 170
#> [20]  22 100 113 107  29  64 103   2 146  67 133  94 152  17  92  66 199 112 134
#> [39]  77  18 176 188  71  41 171 163   1 119  70 108  69 174  48 177  33  78  99
#> [58] 198 129 102  20 135  79 137  11  37
#> 
#> 
#> [[2]]
#> [[2]]$train
#>   [1]   1   2   3   4   5   8  10  11  12  15  17  18  20  21  22  23  25  29
#>  [19]  30  31  32  33  37  40  41  42  43  46  47  48  49  50  53  56  57  58
#>  [37]  59  63  64  66  67  69  70  71  72  73  74  75  77  78  79  82  85  88
#>  [55]  89  90  91  92  93  94  95  98  99 100 101 102 103 104 105 106 107 108
#>  [73] 109 112 113 116 118 119 122 123 124 125 126 127 129 131 132 133 134 135
#>  [91] 136 137 138 139 140 142 144 146 147 149 151 152 156 158 159 160 163 164
#> [109] 168 170 171 172 174 176 177 178 179 181 182 183 185 186 187 188 189 190
#> [127] 192 195 196 197 198 199 200
#> 
#> [[2]]$app
#>  [1]  61  86  35  96  19 184  84 191  36  62  97 180 165  60 173  87  81  44 153
#> [20]   9 162  34 157 154  54  83 155 121 193 148  28  27  13 114  55  24   7  16
#> [39]  52  68 117  26 194  80 130  14 128 115  38 150 169  51 145  65 167 110  45
#> [58] 161   6 141 111 166 175 120  39  76 143
#> 
#> 
#> [[3]]
#> [[3]]$train
#>   [1]   1   2   6   7   9  11  12  13  14  16  17  18  19  20  22  24  26  27
#>  [19]  28  29  32  33  34  35  36  37  38  39  40  41  44  45  47  48  51  52
#>  [37]  54  55  59  60  61  62  64  65  66  67  68  69  70  71  72  73  75  76
#>  [55]  77  78  79  80  81  83  84  86  87  90  92  94  95  96  97  99 100 101
#>  [73] 102 103 107 108 110 111 112 113 114 115 117 119 120 121 125 128 129 130
#>  [91] 131 133 134 135 136 137 138 141 143 145 146 148 150 152 153 154 155 156
#> [109] 157 161 162 163 165 166 167 169 170 171 173 174 175 176 177 180 184 185
#> [127] 187 188 191 193 194 198 199
#> 
#> [[3]]$app
#>  [1] 127 139  42 122 144 179  58  88 189 142  57 158  46   5   4 132  91 183  31
#> [20] 172 118 159  63  53  85 126 106 196 164 186 168  43  21 197  56 160 151 105
#> [39] 123 181 178 149 109 124 140 147  15 182  23  74 190  98   8 200  49 116  30
#> [58] 195  82  25  93  10  89 192  50   3 104
#> 
#> 
#> attr(,"splitmethod")
#> [1] "kwaycross"

# longer example
# helper fns
# fit models using experiment plan to estimate out of sample behavior
fitModelAndApply <- function(trainData,applicaitonData) {
   model <- lm(y~x,data=trainData)
   predict(model,newdata=applicaitonData)
}
simulateOutOfSampleTrainEval <- function(d,fitApplyFn) {
   eSets <- buildEvalSets(nrow(d))
   evals <- lapply(eSets,
      function(ei) { fitApplyFn(d[ei$train,],d[ei$app,]) })
   pred <- numeric(nrow(d))
   for(eii in seq_len(length(eSets))) {
     pred[eSets[[eii]]$app] <- evals[[eii]]
   }
   pred
}

# run the experiment
set.seed(2352356)
# example data
d <- data.frame(x=rnorm(5),y=rnorm(5),
        outOfSampleEst=NA,inSampleEst=NA)

# fit model on all data
d$inSampleEst <- fitModelAndApply(d,d)
# compute in-sample R^2 (above zero, falsely shows a 
#   relation until we adjust for degrees of freedom)
1-sum((d$y-d$inSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] 0.4193942

d$outOfSampleEst <- simulateOutOfSampleTrainEval(d,fitModelAndApply)
# compute out-sample R^2 (not positive, 
#  evidence of no relation)
1-sum((d$y-d$outOfSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] -0.7843929