Build set carve-up for out-of sample evaluation.

Return a carve-up of seq_len(nRows). Very useful for any sort of nested model situation (such as data prep, stacking, or super-learning).

buildEvalSets(nRows, ..., dframe = NULL, y = NULL, splitFunction = NULL,
  nSplits = 3)

Arguments

nRows
scalar, >=1 number of rows to sample from.

...
no additional arguments, declared to forced named binding of later arguments.

dframe
(optional) original data.frame, passed to user splitFunction.

y
(optional) numeric vector, outcome variable (possibly to stratify on), passed to user splitFunction.

splitFunction
(optional) function taking arguments nSplits,nRows,dframe, and y; returning a user desired split.

nSplits
integer, target number of splits.

Value

list of lists where the app portion of the sub-lists is a disjoint carve-up of seq_len(nRows) and each list as a train portion disjoint from app.

Details

Also sets attribute "splitmethod" on return value that describes how the split was performed. attr(returnValue,'splitmethod') is one of: 'notsplit' (data was not split; corner cases like single row data sets), 'oneway' (leave one out holdout), 'kwaycross' (a simple partition), 'userfunction' (user supplied function was actually used), or a user specified attribute. Any user desired properties (such as stratification on y, or preservation of groups designated by original data row numbers) may not apply unless you see that 'userfunction' has been used.

The intent is the user splitFunction only needs to handle "easy cases" and maintain user invariants. If the user splitFunction returns NULL, throws, or returns an unacceptable carve-up then vtreat::buildEvalSets returns its own eval set plan. The signature of splitFunction should be splitFunction(nRows,nSplits,dframe,y) where nSplits is the number of pieces we want in the carve-up, nRows is the number of rows to split, dframe is the original dataframe (useful for any group control variables), and y is a numeric vector representing outcome (useful for outcome stratification).

Note that buildEvalSets may not always return a partition (such as one row dataframes), or if the user split function chooses to make rows eligable for applicaton a different number of times.

See also

Examples

# use
buildEvalSets(200)

#> [[1]]
#> [[1]]$train
#>   [1]   1   2   5   7   8   9  10  12  13  14  15  16  17  19  21  23  24  26
#>  [19]  27  29  30  32  33  34  35  36  37  38  40  41  42  43  44  45  47  49
#>  [37]  50  52  53  54  56  57  58  59  60  62  63  65  66  67  68  69  70  73
#>  [55]  75  76  77  79  80  82  83  84  85  89  90  91  94  95  96  97 100 101
#>  [73] 102 104 106 107 109 111 112 113 114 115 117 118 120 121 123 125 126 128
#>  [91] 130 131 133 136 137 139 141 142 143 146 148 149 151 152 154 155 160 163
#> [109] 164 165 166 167 168 170 172 173 174 175 176 178 179 184 185 188 189 190
#> [127] 191 193 194 195 196 197 198 199
#> 
#> [[1]]$app
#>  [1]  98 140 159  64 177  51   6   4 150 103  81 187 162 135  86  72  61  99  74
#> [20] 124 138 105 158 119 183 161 171 181 180 129 147  25 153  22  55 156 182  87
#> [39] 192  20 200  31  48 132 127  11 108  88 122 110  71  46 169  78  93 144   3
#> [58]  18  39 134 116  92 157 186 145  28
#> 
#> 
#> [[2]]
#> [[2]]$train
#>   [1]   3   4   5   6   8  10  11  12  15  16  18  19  20  22  25  28  29  30
#>  [19]  31  32  34  36  37  38  39  41  43  44  46  47  48  49  50  51  54  55
#>  [37]  61  62  63  64  65  66  68  69  70  71  72  73  74  78  81  83  85  86
#>  [55]  87  88  91  92  93  94  96  98  99 102 103 105 106 107 108 109 110 115
#>  [73] 116 119 121 122 124 126 127 128 129 131 132 133 134 135 136 137 138 139
#>  [91] 140 141 142 144 145 147 149 150 152 153 155 156 157 158 159 160 161 162
#> [109] 163 165 166 168 169 171 172 173 175 177 179 180 181 182 183 184 185 186
#> [127] 187 188 190 192 195 198 200
#> 
#> [[2]]$app
#>  [1]  35 154  13  53 113 101  82  84  67 176  45 197 111  24 114 130 123  33  75
#> [20] 120   2 193  79 146 112  97  21 164  80 143 117 167  95  23 170 151  90  52
#> [39] 148  60   1  58  26   7  77  17 178 191  76  59 199 196  56 125  40  27  14
#> [58] 174   9 100 118  42 194 104  89  57 189
#> 
#> 
#> [[3]]
#> [[3]]$train
#>   [1]   1   2   3   4   6   7   9  11  13  14  17  18  20  21  22  23  24  25
#>  [19]  26  27  28  31  33  35  39  40  42  45  46  48  51  52  53  55  56  57
#>  [37]  58  59  60  61  64  67  71  72  74  75  76  77  78  79  80  81  82  84
#>  [55]  86  87  88  89  90  92  93  95  97  98  99 100 101 103 104 105 108 110
#>  [73] 111 112 113 114 116 117 118 119 120 122 123 124 125 127 129 130 132 134
#>  [91] 135 138 140 143 144 145 146 147 148 150 151 153 154 156 157 158 159 161
#> [109] 162 164 167 169 170 171 174 176 177 178 180 181 182 183 186 187 189 191
#> [127] 192 193 194 196 197 199 200
#> 
#> [[3]]$app
#>  [1] 106  41  69 109  36 185  68  70 198  91  50 173  10  83  38   5 152  62  96
#> [20] 184 141  49  85 149  32  15 168  66  29   8  19  63 133  73 155  44 131 137
#> [39] 190 102 172 126  54 160 128  94 121 136  43 188 195 163  47  12  37 115 107
#> [58]  65 179 165 166 175 139  34 142  30  16
#> 
#> 
#> attr(,"splitmethod")
#> [1] "kwaycross"

# longer example
# helper fns
# fit models using experiment plan to estimate out of sample behavior
fitModelAndApply <- function(trainData,applicaitonData) {
   model <- lm(y~x,data=trainData)
   predict(model,newdata=applicaitonData)
}
simulateOutOfSampleTrainEval <- function(d,fitApplyFn) {
   eSets <- buildEvalSets(nrow(d))
   evals <- lapply(eSets,
      function(ei) { fitApplyFn(d[ei$train,],d[ei$app,]) })
   pred <- numeric(nrow(d))
   for(eii in seq_len(length(eSets))) {
     pred[eSets[[eii]]$app] <- evals[[eii]]
   }
   pred
}

# run the experiment
set.seed(2352356)
# example data
d <- data.frame(x=rnorm(5),y=rnorm(5),
        outOfSampleEst=NA,inSampleEst=NA)

# fit model on all data
d$inSampleEst <- fitModelAndApply(d,d)
# compute in-sample R^2 (above zero, falsely shows a 
#   relation until we adjust for degrees of freedom)
1-sum((d$y-d$inSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] 0.4193942

d$outOfSampleEst <- simulateOutOfSampleTrainEval(d,fitModelAndApply)
# compute out-sample R^2 (not positive, 
#  evidence of no relation)
1-sum((d$y-d$outOfSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] -0.7843929