Build set carve-up for out-of sample evaluation.

Return a carve-up of seq_len(nRows). Very useful for any sort of nested model situation (such as data prep, stacking, or super-learning).

buildEvalSets(nRows, ..., dframe = NULL, y = NULL, splitFunction = NULL,
  nSplits = 3)

Arguments

nRows
scalar, >=1 number of rows to sample from.

...
no additional arguments, declared to forced named binding of later arguments.

dframe
(optional) original data.frame, passed to user splitFunction.

y
(optional) numeric vector, outcome variable (possibly to stratify on), passed to user splitFunction.

splitFunction
(optional) function taking arguments nSplits,nRows,dframe, and y; returning a user desired split.

nSplits
integer, target number of splits.

Value

list of lists where the app portion of the sub-lists is a disjoint carve-up of seq_len(nRows) and each list as a train portion disjoint from app.

Details

Also sets attribute "splitmethod" on return value that describes how the split was performed. attr(returnValue,'splitmethod') is one of: 'notsplit' (data was not split; corner cases like single row data sets), 'oneway' (leave one out holdout), 'kwaycross' (a simple partition), 'userfunction' (user supplied function was actually used), or a user specified attribute. Any user desired properties (such as stratification on y, or preservation of groups designated by original data row numbers) may not apply unless you see that 'userfunction' has been used.

The intent is the user splitFunction only needs to handle "easy cases" and maintain user invariants. If the user splitFunction returns NULL, throws, or returns an unacceptable carve-up then vtreat::buildEvalSets returns its own eval set plan. The signature of splitFunction should be splitFunction(nRows,nSplits,dframe,y) where nSplits is the number of pieces we want in the carve-up, nRows is the number of rows to split, dframe is the original dataframe (useful for any group control variables), and y is a numeric vector representing outcome (useful for outcome stratification).

Note that buildEvalSets may not always return a partition (such as one row dataframes), or if the user split function chooses to make rows eligable for applicaton a different number of times.

See also

kWayCrossValidation, kWayStratifiedY, and makekWayCrossValidationGroupedByColumn

Examples

# use
buildEvalSets(200)

#> [[1]]
#> [[1]]$train
#>   [1]   1   2   3   4   6   7  10  11  13  14  17  18  21  22  23  25  26  27
#>  [19]  29  31  36  37  38  39  40  43  44  45  46  47  49  54  56  58  59  60
#>  [37]  61  63  65  66  68  71  73  76  77  78  80  81  82  83  84  85  86  88
#>  [55]  92  93  94  96  97  98  99 101 102 103 104 105 106 107 108 109 110 111
#>  [73] 113 115 116 117 120 121 122 123 124 128 129 130 131 132 133 134 136 138
#>  [91] 139 140 143 144 145 146 148 149 150 151 155 157 158 161 163 164 167 168
#> [109] 169 170 172 173 175 176 177 179 180 181 182 183 184 185 186 187 190 191
#> [127] 193 194 195 196 197 198 199 200
#> 
#> [[1]]$app
#>  [1] 119  91 141  34  75  12  53  35   9 118  51  72  30 166  15 152  70  79 162
#> [20]  50 147 153 165  48 192  69  64 189 137 188  57 135 125 100  41  62  95  89
#> [39] 171  52  20  19 174 114 178  16   5 142  33  42 126 156  87  32  28   8  90
#> [58]  74 160  67 159  24 154  55 127 112
#> 
#> 
#> [[2]]
#> [[2]]$train
#>   [1]   1   2   5   8   9  10  11  12  15  16  18  19  20  21  22  23  24  26
#>  [19]  28  30  32  33  34  35  36  39  41  42  43  44  45  48  50  51  52  53
#>  [37]  54  55  56  57  59  60  61  62  63  64  66  67  69  70  72  74  75  76
#>  [55]  77  79  80  81  82  84  85  86  87  88  89  90  91  93  95 100 101 104
#>  [73] 106 109 112 113 114 115 118 119 120 121 124 125 126 127 130 131 134 135
#>  [91] 136 137 139 140 141 142 145 146 147 149 151 152 153 154 155 156 157 159
#> [109] 160 162 163 165 166 167 169 171 174 176 177 178 180 183 184 186 187 188
#> [127] 189 192 194 196 197 199 200
#> 
#> [[2]]$app
#>  [1]  17  31  97 148   7  37  71 122 173  92   6 107 116 181 102 198 191  47  73
#> [20] 103 164  46 128 185   4   3 105 132 193  25 110 195 108 129  40 172  94  78
#> [39] 143  65 123 182 144  14 175  38 158  68 168  13 161  49 179  83  99  96 133
#> [58] 111 150  98 190 117  29 138  27  58 170
#> 
#> 
#> [[3]]
#> [[3]]$train
#>   [1]   3   4   5   6   7   8   9  12  13  14  15  16  17  19  20  24  25  27
#>  [19]  28  29  30  31  32  33  34  35  37  38  40  41  42  46  47  48  49  50
#>  [37]  51  52  53  55  57  58  62  64  65  67  68  69  70  71  72  73  74  75
#>  [55]  78  79  83  87  89  90  91  92  94  95  96  97  98  99 100 102 103 105
#>  [73] 107 108 110 111 112 114 116 117 118 119 122 123 125 126 127 128 129 132
#>  [91] 133 135 137 138 141 142 143 144 147 148 150 152 153 154 156 158 159 160
#> [109] 161 162 164 165 166 168 170 171 172 173 174 175 178 179 181 182 185 188
#> [127] 189 190 191 192 193 195 198
#> 
#> [[3]]$app
#>  [1] 167   2  56 199  60 186 177 131 130 120  39  80 155 109 104 163  59  26 113
#> [20]  10 169  82 146 151 187 149 124 197 106  76 200  77  86  81  22 101 115 196
#> [39]  43 176  45  66 157 183  63  44 140  36  93 136  11  21 121 139 194 145  23
#> [58]   1 184  54  88  61 134  84  18 180  85
#> 
#> 
#> attr(,"splitmethod")
#> [1] "kwaycross"

# longer example
# helper fns
# fit models using experiment plan to estimate out of sample behavior
fitModelAndApply <- function(trainData,applicaitonData) {
   model <- lm(y~x,data=trainData)
   predict(model,newdata=applicaitonData)
}
simulateOutOfSampleTrainEval <- function(d,fitApplyFn) {
   eSets <- buildEvalSets(nrow(d))
   evals <- lapply(eSets,
      function(ei) { fitApplyFn(d[ei$train,],d[ei$app,]) })
   pred <- numeric(nrow(d))
   for(eii in seq_len(length(eSets))) {
     pred[eSets[[eii]]$app] <- evals[[eii]]
   }
   pred
}

# run the experiment
set.seed(2352356)
# example data
d <- data.frame(x=rnorm(5),y=rnorm(5),
        outOfSampleEst=NA,inSampleEst=NA)

# fit model on all data
d$inSampleEst <- fitModelAndApply(d,d)
# compute in-sample R^2 (above zero, falsely shows a 
#   relation until we adjust for degrees of freedom)
1-sum((d$y-d$inSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] 0.4193942

d$outOfSampleEst <- simulateOutOfSampleTrainEval(d,fitModelAndApply)
# compute out-sample R^2 (not positive, 
#  evidence of no relation)
1-sum((d$y-d$outOfSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] -0.568004