Build set carve-up for out-of sample evaluation.

Return a carve-up of seq_len(nRows). Very useful for any sort of nested model situation (such as data prep, stacking, or super-learning).

buildEvalSets(nRows, ..., dframe = NULL, y = NULL, splitFunction = NULL,
  nSplits = 3)

Arguments

nRows
scalar, >=1 number of rows to sample from.

...
no additional arguments, declared to forced named binding of later arguments.

dframe
(optional) original data.frame, passed to user splitFunction.

y
(optional) numeric vector, outcome variable (possibly to stratify on), passed to user splitFunction.

splitFunction
(optional) function taking arguments nSplits,nRows,dframe, and y; returning a user desired split.

nSplits
integer, target number of splits.

Value

list of lists where the app portion of the sub-lists is a disjoint carve-up of seq_len(nRows) and each list as a train portion disjoint from app.

Details

Also sets attribute "splitmethod" on return value that describes how the split was performed. attr(returnValue,'splitmethod') is one of: 'notsplit' (data was not split; corner cases like single row data sets), 'oneway' (leave one out holdout), 'kwaycross' (a simple partition), 'userfunction' (user supplied function was actually used), or a user specified attribute. Any user desired properties (such as stratification on y, or preservation of groups designated by original data row numbers) may not apply unless you see that 'userfunction' has been used.

The intent is the user splitFunction only needs to handle "easy cases" and maintain user invariants. If the user splitFunction returns NULL, throws, or returns an unacceptable carve-up then vtreat::buildEvalSets returns its own eval set plan. The signature of splitFunction should be splitFunction(nRows,nSplits,dframe,y) where nSplits is the number of pieces we want in the carve-up, nRows is the number of rows to split, dframe is the original dataframe (useful for any group control variables), and y is a numeric vector representing outcome (useful for outcome stratification).

Note that buildEvalSets may not always return a partition (such as one row dataframes), or if the user split function chooses to make rows eligable for applicaton a different number of times.

See also

Examples

# use
buildEvalSets(200)

#> [[1]]
#> [[1]]$train
#>   [1]   1   2   4   6   7   8   9  12  13  14  16  17  22  24  25  26  27  28
#>  [19]  29  30  31  32  33  35  36  37  40  41  42  44  48  49  50  52  54  59
#>  [37]  61  62  63  64  66  68  69  70  72  73  74  75  76  79  80  81  82  83
#>  [55]  84  85  86  87  91  92  93  95  97  99 100 103 104 105 106 107 111 112
#>  [73] 113 115 116 117 120 121 122 124 125 126 127 128 129 130 131 133 134 135
#>  [91] 137 139 141 143 145 147 148 149 150 151 152 153 156 157 158 159 160 161
#> [109] 162 163 165 166 167 168 171 173 174 175 177 178 179 181 182 183 184 185
#> [127] 187 188 190 194 195 197 199 200
#> 
#> [[1]]$app
#>  [1]  78 110 102  55 193  11  94  43   5 140  71 109 191 142  56 176  38 169  18
#> [20] 180  77  34  10  23 186 101  88  39  53 154 164 172  89 136  65 170 198 132
#> [39]  58 155  96 108 119   3  15  60 144  57  46 138 196  45 189 118 146  67  21
#> [58]  19  90 114  51  47  98 192  20 123
#> 
#> 
#> [[2]]
#> [[2]]$train
#>   [1]   1   2   3   5   8  10  11  12  13  15  16  18  19  20  21  22  23  24
#>  [19]  27  33  34  37  38  39  43  45  46  47  48  50  51  53  54  55  56  57
#>  [37]  58  60  63  64  65  67  71  72  73  74  75  76  77  78  79  81  85  86
#>  [55]  88  89  90  92  93  94  96  98 101 102 103 105 106 108 109 110 113 114
#>  [73] 118 119 122 123 124 125 126 128 129 132 133 135 136 137 138 139 140 142
#>  [91] 144 145 146 147 148 149 150 153 154 155 157 159 161 162 164 165 166 168
#> [109] 169 170 171 172 173 174 175 176 180 181 185 186 187 188 189 190 191 192
#> [127] 193 194 195 196 197 198 199
#> 
#> [[2]]$app
#>  [1]   6   4 167 115  91  40 182 152  97  82  69 112  84 141 158 121 131 163  59
#> [20]  28 120  80 200 179 117  30 151  26  66 104 100 107  31  25 177 178  62   7
#> [39]  42  14  29 111 130  68  87  83  70  52  99  41 127 184  32   9 160 116  44
#> [58] 183  49 134  35  17 156  61  95  36 143
#> 
#> 
#> [[3]]
#> [[3]]$train
#>   [1]   3   4   5   6   7   9  10  11  14  15  17  18  19  20  21  23  25  26
#>  [19]  28  29  30  31  32  34  35  36  38  39  40  41  42  43  44  45  46  47
#>  [37]  49  51  52  53  55  56  57  58  59  60  61  62  65  66  67  68  69  70
#>  [55]  71  77  78  80  82  83  84  87  88  89  90  91  94  95  96  97  98  99
#>  [73] 100 101 102 104 107 108 109 110 111 112 114 115 116 117 118 119 120 121
#>  [91] 123 127 130 131 132 134 136 138 140 141 142 143 144 146 151 152 154 155
#> [109] 156 158 160 163 164 167 169 170 172 176 177 178 179 180 182 183 184 186
#> [127] 189 191 192 193 196 198 200
#> 
#> [[3]]$app
#>  [1]  93  74  75  50 173 124  27 128 147 139  37  85 137   2  79 188  86 129 194
#> [20]  24 148 161 122 149 174 113 181  54 125 165  64  76 126   1  73 135   8 106
#> [39]  22 157 168 199 153  16 103 171  48  63 133 166  33  92 150  81 105 162 195
#> [58]  72 185  12  13 190 145 159 197 175 187
#> 
#> 
#> attr(,"splitmethod")
#> [1] "kwaycross"

# longer example
# helper fns
# fit models using experiment plan to estimate out of sample behavior
fitModelAndApply <- function(trainData,applicaitonData) {
   model <- lm(y~x,data=trainData)
   predict(model,newdata=applicaitonData)
}
simulateOutOfSampleTrainEval <- function(d,fitApplyFn) {
   eSets <- buildEvalSets(nrow(d))
   evals <- lapply(eSets,
      function(ei) { fitApplyFn(d[ei$train,],d[ei$app,]) })
   pred <- numeric(nrow(d))
   for(eii in seq_len(length(eSets))) {
     pred[eSets[[eii]]$app] <- evals[[eii]]
   }
   pred
}

# run the experiment
set.seed(2352356)
# example data
d <- data.frame(x=rnorm(5),y=rnorm(5),
        outOfSampleEst=NA,inSampleEst=NA)

# fit model on all data
d$inSampleEst <- fitModelAndApply(d,d)
# compute in-sample R^2 (above zero, falsely shows a 
#   relation until we adjust for degrees of freedom)
1-sum((d$y-d$inSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] 0.4193942

d$outOfSampleEst <- simulateOutOfSampleTrainEval(d,fitModelAndApply)
# compute out-sample R^2 (not positive, 
#  evidence of no relation)
1-sum((d$y-d$outOfSampleEst)^2)/sum((d$y-mean(d$y))^2)

#> [1] -0.7843929