Function to apply mkCrossFrameMExperiment treatemnts.

Please see vignette("MultiClassVtreat", package = "vtreat") https://winvector.github.io/vtreat/articles/MultiClassVtreat.html.

# S3 method for multinomial_plan
prepare(
  treatmentplan,
  dframe,
  ...,
  pruneSig = NULL,
  scale = FALSE,
  doCollar = FALSE,
  varRestriction = NULL,
  codeRestriction = NULL,
  trackedValues = NULL,
  extracols = NULL,
  parallelCluster = NULL,
  use_parallel = TRUE
)

Arguments

treatmentplan
multinomial_plan from mkCrossFrameMExperiment.

dframe
new data to process.

...
not used, declared to forced named binding of later arguments

pruneSig
suppress variables with significance above this level

scale
optional if TRUE replace numeric variables with single variable model regressions ("move to outcome-scale"). These have mean zero and (for variables with significant less than 1) slope 1 when regressed (lm for regression problems/glm for classification problems) against outcome.

doCollar
optional if TRUE collar numeric variables by cutting off after a tail-probability specified by collarProb during treatment design.

varRestriction
optional list of treated variable names to restrict to

codeRestriction
optional list of treated variable codes to restrict to

trackedValues
optional named list mapping variables to know values, allows warnings upon novel level appearances (see track_values)

extracols
extra columns to copy.

parallelCluster
(optional) a cluster object created by package parallel or package snow.

use_parallel
logical, if TRUE use parallel methods.

Value

prepared data frame.

See also