Apply treatments and restrict to useful variables.

Use a treatment plan to prepare a data frame for analysis. The resulting frame will have new effective variables that are numeric and free of NaN/NA. If the outcome column is present it will be copied over. The intent is that these frames are compatible with more machine learning techniques, and avoid a lot of corner cases (NA,NaN, novel levels, too many levels). Note: each column is processed independently of all others. Also copies over outcome if present. Note: treatmentplan's are not meant for long-term storage, a warning is issued if the version of vtreat that produced the plan differs from the version running prepare().

# S3 method for treatmentplan
prepare(treatmentplan, dframe, ...,
  pruneSig = NULL, scale = FALSE, doCollar = FALSE,
  varRestriction = NULL, codeRestriction = NULL,
  trackedValues = NULL, extracols = NULL, parallelCluster = NULL,
  use_parallel = TRUE)

Arguments

treatmentplan
Plan built by designTreantmentsC() or designTreatmentsN()

dframe
Data frame to be treated

...
no additional arguments, declared to forced named binding of later arguments

pruneSig
suppress variables with significance above this level

scale
optional if TRUE replace numeric variables with single variable model regressions ("move to outcome-scale"). These have mean zero and (for variables with significant less than 1) slope 1 when regressed (lm for regression problems/glm for classification problems) against outcome.

doCollar
optional if TRUE collar numeric variables by cutting off after a tail-probability specified by collarProb during treatment design.

varRestriction
optional list of treated variable names to restrict to

codeRestriction
optional list of treated variable codes to restrict to

trackedValues
optional named list mapping variables to know values, allows warnings upon novel level appearances (see track_values)

extracols
extra columns to copy.

parallelCluster
(optional) a cluster object created by package parallel or package snow.

use_parallel
logical, if TRUE use parallel methods.

Value

treated data frame (all columns numeric- without NA, NaN)

See also

Examples

dTrainN <- data.frame(x= c('a','a','a','a','b','b','b'),
                      z= c(1,2,3,4,5,6,7),
                      y= c(0,0,0,1,0,1,1))
dTestN <- data.frame(x= c('a','b','c',NA),
                     z= c(10,20,30,NA))
treatmentsN = designTreatmentsN(dTrainN,colnames(dTrainN), 'y')

#> [1] "vtreat 1.3.6 inspecting inputs Thu Feb  7 10:46:59 2019"
#> [1] "designing treatments Thu Feb  7 10:46:59 2019"
#> [1] " have initial level statistics Thu Feb  7 10:46:59 2019"
#> [1] " scoring treatments Thu Feb  7 10:46:59 2019"
#> [1] "have treatment plan Thu Feb  7 10:46:59 2019"

dTrainNTreated <- prepare(treatmentsN, dTrainN, pruneSig= 0.2)
dTestNTreated <- prepare(treatmentsN, dTestN, pruneSig= 0.2)

dTrainC <- data.frame(x= c('a','a','a','b','b','b'),
                      z= c(1,2,3,4,5,6),
                      y= c(FALSE,FALSE,TRUE,FALSE,TRUE,TRUE))
dTestC <- data.frame(x= c('a','b','c',NA),
                     z= c(10,20,30,NA))
treatmentsC <- designTreatmentsC(dTrainC, colnames(dTrainC),'y',TRUE)

#> [1] "vtreat 1.3.6 inspecting inputs Thu Feb  7 10:46:59 2019"
#> [1] "designing treatments Thu Feb  7 10:46:59 2019"
#> [1] " have initial level statistics Thu Feb  7 10:46:59 2019"
#> [1] " scoring treatments Thu Feb  7 10:46:59 2019"
#> [1] "have treatment plan Thu Feb  7 10:46:59 2019"

dTrainCTreated <- prepare(treatmentsC, dTrainC)
dTestCTreated <- prepare(treatmentsC, dTestC)

dTrainZ <- data.frame(x= c('a','a','a','b','b','b'),
                      z= c(1,2,3,4,5,6))
dTestZ <- data.frame(x= c('a','b','c',NA),
                     z= c(10,20,30,NA))
treatmentsZ <- designTreatmentsZ(dTrainZ, colnames(dTrainZ))

#> [1] "vtreat 1.3.6 inspecting inputs Thu Feb  7 10:46:59 2019"
#> [1] "designing treatments Thu Feb  7 10:46:59 2019"
#> [1] " have initial level statistics Thu Feb  7 10:46:59 2019"
#> [1] " scoring treatments Thu Feb  7 10:46:59 2019"
#> [1] "have treatment plan Thu Feb  7 10:46:59 2019"

dTrainZTreated <- prepare(treatmentsZ, dTrainZ, codeRestriction= c('lev'))
dTestZTreated <- prepare(treatmentsZ, dTestZ, codeRestriction= c('lev'))